+++ game-frontend/CHANGELOG.md
@@ -3,6 +3,7 @@
 v0.10.0 (latest)
 =================
 * Merge v0.9.21
+* __Feature:__ [As a player I want to have invisible units #11](https://github.com/KevinGuancheDarias/owge/issues/11)  
 * __Feature:__ [As a player I want to have units capable of bypassing the unit shield #270](https://github.com/KevinGuancheDarias/owge/issues/270)  
 * [class=Admin] __Feature:__ [As an admin I want to be able to define units that can intercept units with a given speed group #21](https://github.com/KevinGuancheDarias/owge/issues/21)  
 * [class=Admin] __Feature:__ [As an Admin I want to be able to override the maxCount of a certain unit type for a given faction #273](https://github.com/KevinGuancheDarias/owge/issues/273)